requirements.txt
cycler==0.10.0
decorator==4.4.2
future==0.18.2
h5py==2.10.0
imageio==2.8.0
networkx==2.4
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.15.0
scipy==1.4.1
six==1.14.0
tifffile==2019.7.26.2
torch==1.5.0
